update check id before buy

+++ app/repos/cart.py
@@ -1,4 +1,4 @@
-from sqlalchemy.orm import Session
+from sqlalchemy.orm import Session,load_only
 from .. import tables
 from fastapi import HTTPException,status
 
@@ -60,28 +60,34 @@ def update_quantity(request, db: Session,user):
 
 def buy_item(ids,db: Session,user):
     items=db.query(tables.CartItems).filter(tables.CartItems.user_id==user.id)
-    new_order=tables.Order(user_id=user.id)
-    db.add(new_order)
-    db.commit()
-    db.refresh(new_order)
-    for id in ids:
-        #add to order
-        quantity=item.product_quantity
-        price=item.prd_inf.price
-        item=items.filter(tables.CartItems.product_id==id).first()
-        order_detail=tables.OrderDetail(order_id=new_order.id,
-                                        product_id=id,
-                                        quantity=quantity,
-                                        price=price
-                                        )
-        db.add(order_detail)
+    ids_inDB=[item.product_id for item in items.all()]
+    check =  all(id in ids_inDB for id in ids)
+    if check:
+        new_order=tables.Order(user_id=user.id)
+        db.add(new_order)
         db.commit()
-        db.refresh(order_detail)
-        #update quantity,sold in storage
-        db.query(tables.product).filter(tables.product.id==id).update({tables.product.quantity:tables.product.quantity-item.product_quantity})
-        db.query(tables.product).filter(tables.product.id==id).update({tables.product.sold:tables.product.sold+item.product_quantity})
-        #delete from cart
-        db.delete(item)
-        db.commit()
+        db.refresh(new_order)
+        for id in ids:
+            #add to order
+            item=items.filter(tables.CartItems.product_id==id).first()
+            quantity=item.product_quantity
+            price=item.prd_inf.price
+            order_detail=tables.OrderDetail(order_id=new_order.id,
+                                            product_id=id,
+                                            quantity=quantity,
+                                            price=price
+                                            )
+            db.add(order_detail)
+            db.commit()
+            db.refresh(order_detail)
+            #update quantity,sold in storage
+            db.query(tables.product).filter(tables.product.id==id).update({tables.product.quantity:tables.product.quantity-item.product_quantity})
+            db.query(tables.product).filter(tables.product.id==id).update({tables.product.sold:tables.product.sold+item.product_quantity})
+            #delete from cart
+            db.delete(item)
+            db.commit()
+    else:
+         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                        detail=f"Item doesn't exist")
     return {"message":"Buy successfully",
             "order detail":new_order.order_detail}
